backend/requirements.txt
flask
flask-cors
requests
beautifulsoup4
sentence-transformers
qdrant-client
google-generativeai
